+++ BlindBox Backend/src/main/java/com/blindbox/service/impl/ProductServiceImpl.java
@@ -41,7 +41,11 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void deleteProduct(Integer id) {
-        productRepository.deleteById(id);
+    public boolean deleteProduct(Integer id) {
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
