+++ BlindBox Backend/src/main/java/com/blindbox/service/impl/ReviewServiceImpl.java
@@ -21,21 +21,6 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public List<Review> getReviewsByUserId(Integer userId) {
-        return reviewRepository.findByUserId(userId);
-    }
-
-    @Override
-    public List<Review> getReviewsByBlinkboxId(Integer blinkboxId) {
-        return reviewRepository.findByBlindboxId(blinkboxId);
-    }
-
-    @Override
-    public List<Review> getReviewsByProductId(Integer productId) {
-        return reviewRepository.findByProductId(productId);
-    }
-
-    @Override
     public Review getReviewById(Integer id) {
         Optional<Review> review = reviewRepository.findById(id);
         return review.orElse(null);
@@ -47,7 +32,20 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public void deleteReview(Integer id) {
-        reviewRepository.deleteById(id);
+    public Review updateReview(Integer id, Review review) {
+        if (reviewRepository.existsById(id)) {
+            review.setReviewID(id);
+            return reviewRepository.save(review);
+        }
+        return null;
+    }
+
+    @Override
+    public boolean deleteReview(Integer id) {
+        if (reviewRepository.existsById(id)) {
+            reviewRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
